+++ NEWS.md
@@ -1,5 +1,7 @@
 # NEWS rtide
 
+- Improved harmonics data checking.
+
 #### v0.0.1
 
 - Initial Release